requirements.txt
Django==4.2.7
django-cors-headers==4.3.1
Pillow==10.4.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
gunicorn==21.2.0
python-decouple==3.8
dj-database-url==2.1.0
python-dotenv==1.0.0
django-crispy-forms==2.1
crispy-bootstrap5==0.7
markdown==3.5.2
pdfkit==1.0.0
xlsxwriter==3.1.9
xlwt==1.3.0
pywebpush==2.0.3
openpyxl==3.1.2
djangorestframework==3.14.0
weasyprint==60.2
